Update to use Gloss 1.10
The gloss pull request benl23x5/gloss#21 changes the API in a way that
allows us to skip the color channel and row re-ordering.  This is
significant because we no longer need to perform an O(N) copy of every
image, and instead propogate the row and pixel information to the
underlying OpenGL system (which can then do any number of things, but
whatever it does will probably be better optimized).